fix(rebase): resolve thermal-manifold conflict with #1474
PR #1474 added the real ThermalManifold to src/physics/geometry_tensor.rs
while #1473's branch (Phase 1b) was already merged into a stub named
ThermalManifold to bridge against the not-yet-merged Phase 1a data
structure. Git's three-way merge combined both into the file, leaving two
conflicting ThermalManifold type definitions that broke compilation.

Resolution:
- Drop the stub ThermalManifold + GAUGE_CONNECTION_* constants from
  geometry_tensor.rs (the real Phase 1a definitions from #1474 remain).
- Move the stub into gauge_solver.rs as a module-private type used only
  by GaugeSolver's shadow-mode bookkeeping (it never leaves the module).
- Demote GaugeSolver::new / GaugeSolver::manifold() to pub(crate) since
  the private ThermalManifold cannot be exposed across the public
  boundary; default() and the in-module tests still construct /
  inspect it cleanly.

No behavior change to GaugeSolver shadow mode — public API
(GaugeBoundaryConditions, translate_boundary_conditions,
effective_exterior_temperature, step_with_boundary_conditions,
HeatConductionSolver) is byte-identical.
