+++ src/physics/geometry_tensor.rs
@@ -247,79 +247,6 @@ impl GeometryTensor {
     }
 }
 
-/// Phase 1b placeholder for #1461's full ThermalManifold. This stub only carries per-zone gauge_connection boundary vectors so GaugeSolver shadow mode can compile until Phase 1a replaces it.
-#[derive(Debug, Clone)]
-pub struct ThermalManifold {
-    pub geometry: GeometryTensor,
-    pub gauge_connection: Vec<VectorField>,
-}
-
-pub const GAUGE_CONNECTION_COMPONENTS: usize = 2;
-pub const GAUGE_CONNECTION_SOLAR_INDEX: usize = 0;
-pub const GAUGE_CONNECTION_OUTDOOR_TEMP_INDEX: usize = 1;
-
-impl ThermalManifold {
-    pub fn new(num_zones: usize) -> Self {
-        assert!(
-            num_zones <= MAX_ZONES,
-            "ThermalManifold zone count exceeds MAX_ZONES"
-        );
-        let mut geometry = GeometryTensor::new();
-        geometry.summary[0] = num_zones as f64;
-        let gauge_connection = (0..num_zones)
-            .map(|_| VectorField::from_scalar(0.0, GAUGE_CONNECTION_COMPONENTS))
-            .collect();
-
-        Self {
-            geometry,
-            gauge_connection,
-        }
-    }
-
-    pub fn from_geometry(geometry: GeometryTensor) -> Self {
-        let num_zones = geometry.num_zones();
-        assert!(
-            num_zones <= MAX_ZONES,
-            "ThermalManifold zone count exceeds MAX_ZONES"
-        );
-        let gauge_connection = (0..num_zones)
-            .map(|_| VectorField::from_scalar(0.0, GAUGE_CONNECTION_COMPONENTS))
-            .collect();
-
-        Self {
-            geometry,
-            gauge_connection,
-        }
-    }
-
-    pub fn num_zones(&self) -> usize {
-        self.gauge_connection.len()
-    }
-
-    pub fn set_gauge_connection(
-        &mut self,
-        zone_index: usize,
-        solar_irradiance_wm2: f64,
-        outside_air_temp_c: f64,
-    ) -> Result<(), String> {
-        if zone_index >= self.gauge_connection.len() {
-            return Err(format!(
-                "zone_index {} out of bounds for {} zones",
-                zone_index,
-                self.gauge_connection.len()
-            ));
-        }
-
-        self.gauge_connection[zone_index] =
-            VectorField::new(vec![solar_irradiance_wm2, outside_air_temp_c]);
-        Ok(())
-    }
-
-    pub fn gauge_connection(&self, zone_index: usize) -> Option<&VectorField> {
-        self.gauge_connection.get(zone_index)
-    }
-}
-
 /// Wall data structure.
 #[derive(Debug, Clone, Copy)]
 pub struct WallData {
